added bare-bones profile to home.svelte
Added profile page to home.svelte, it is currently empty besides a settings menu, however, it is functional.  on click of profile it will shown = true, then shown = false on second click

Also wrapped home.svelte in a <main> tag which we're using flexbox on to format pages more consistently.