Settings page created
Created bare bones of settings page, this should be the only place that we call the logout functionality, so I also changed the styling of that to simply be text.

The logout button is functional, but the rest of the settings need to be linked up in order to function. We may also in future need to find a better way of getting the "Select Home" info and the "Select Language" info.